+++ pages/marktplatz_liste.inc.php
@@ -51,9 +51,12 @@ $offset = verifyOffset($offset, $offerCount, Config::getInt(Config::SECTION_BASE
             <td>
                 <?php
                 if ($row['VonId'] != $_SESSION['blm_user']) {
-                    echo '<a href="./actions/marktplatz.php?a=2&amp;id=' . $row['ID'] . '&amp;token=' . $_SESSION['blm_xsrf_token'] . '" onclick="return confirm(\'Wollen Sie das Angebot Nr ' . $rowNr . ' wirklich kaufen?\')">Kaufen</a>';
+                    printf('<a class="buy_offer" data-id="%d"  href="./actions/marktplatz.php?a=2&amp;id=%s&amp;token=%s">Kaufen</a>',
+                        $rowNr, $row['ID'], $_SESSION['blm_xsrf_token']);
                 } else {
-                    echo '<a href="./actions/marktplatz.php?a=3&amp;id=' . $row['ID'] . '&amp;token=' . $_SESSION['blm_xsrf_token'] . '" onclick="return confirm(\'Wollen Sie das Angebot Nr ' . $rowNr . ' zurückziehen?\nSie erhalten lediglich ' . formatWeight(floor($row['Menge'] * Config::getFloat(Config::SECTION_MARKET, 'retract_rate'))) . ' der Waren zurück.\')">Zurückziehen</a>';
+                    $refundWeight = formatWeight(floor($row['Menge'] * Config::getFloat(Config::SECTION_MARKET, 'retract_rate')), false);
+                    printf('<a class="retract_offer" data-id="%d" data-refund="%s" href="./actions/marktplatz.php?a=3&amp;id=%s&amp;token=%s">Zurückziehen</a>',
+                        $rowNr, $refundWeight, $row['ID'], $_SESSION['blm_xsrf_token']);
                 }
                 ?>
             </td>
@@ -61,7 +64,7 @@ $offset = verifyOffset($offset, $offerCount, Config::getInt(Config::SECTION_BASE
         <?php
     }
     if (count($entries) == 0) {
-        echo '<tr><td colspan="6" style="text-align: center;"><i>Es wurden keine Angebote gefunden</i></td></tr>';
+        echo '<tr><td colspan="6" class="center"><i>Es wurden keine Angebote gefunden</i></td></tr>';
     }
     ?>
 </table>
@@ -69,3 +72,17 @@ $offset = verifyOffset($offset, $offerCount, Config::getInt(Config::SECTION_BASE
 <?= createPaginationTable('/?p=marktplatz_liste', $offset, $offerCount, Config::getInt(Config::SECTION_BASE, 'market_page_size')); ?>
 
 <a href="/?p=marktplatz_verkaufen">Neues Angebot einstellen</a>
+
+<script nonce="<?= getCspNonce(); ?>">
+    // require confirmation when buying offer
+    for (let buyLink of document.getElementsByClassName('buy_offer')) {
+        let number = buyLink.getAttribute('data-id');
+        buyLink.onclick = () => confirm('Wollen Sie das Angebot Nr ' + number + ' wirklich kaufen?');
+    }
+    // require confirmation when buying offer
+    for (let retractLink of document.getElementsByClassName('retract_offer')) {
+        let number = retractLink.getAttribute('data-id');
+        let refund = retractLink.getAttribute('data-refund');
+        retractLink.onclick = () => confirm('Wollen Sie das Angebot Nr ' + number + ' wirklich zurückziehen?\nSie erhalten lediglich ' + refund + ' kg der Waren zurück.');
+    }
+</script>